refactor(frontend): migrate App.tsx monolith to feature-sliced architecture
