+++ frontend/src/features/residents/components/WebcamFeed.tsx
