+++ frontend/src/pages/AccessPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { AccessGate } from '../features/access/components/AccessGate';
+
+export default function AccessPage({ onRegisterVisitor }: { onRegisterVisitor: () => void }) {
+  return <AccessGate onRegisterVisitor={onRegisterVisitor} />;
+}
+
