+++ frontend/src/pages/DashboardPage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Dashboard } from '../features/alerts/components/Dashboard';
+import { AlertCounts } from '../types';
+
+export default function DashboardPage({ alertCounts }: { alertCounts: AlertCounts }) {
+  return <Dashboard alertCounts={alertCounts} />;
+}
+
