+++ frontend/src/pages/LiveMonitoringPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { LiveMonitoring } from '../features/live-monitoring/components/LiveMonitoring';
+
+export default function LiveMonitoringPage() {
+  return <LiveMonitoring />;
+}
+
