+++ frontend/src/pages/ResidentsPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { Residents } from '../features/residents/components/Residents';
+
+export default function ResidentsPage({ query = '' }: { query?: string }) {
+  return <Residents query={query} />;
+}
+
