+++ frontend/src/pages/SettingsPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { SettingsPanel } from '../features/settings/components/SettingsPanel';
+
+export default function SettingsPage() {
+  return <SettingsPanel />;
+}
+
